refactored into a lib folder
